Move s3 and file_utils under utils folder
